third_party/WebKit/LayoutTests/platform/mac-retina/external/wpt/IndexedDB/parallel-cursors-upgrade-expected.txt

This is a testharness.js-based test.
FAIL 2 cursors assert_equals: Expected success event, but got error event instead expected "success" but got "error"
FAIL 10 cursors assert_equals: Expected success event, but got error event instead expected "success" but got "error"
FAIL 100 cursors assert_equals: Expected success event, but got error event instead expected "success" but got "error"
FAIL 1000 cursors assert_equals: Expected success event, but got error event instead expected "success" but got "error"
FAIL 10000 cursors assert_equals: Expected success event, but got error event instead expected "success" but got "error"
Harness: the test ran to completion.
